+++ lib/minting/money_attribute/configuration.rb
@@ -3,12 +3,11 @@
 module Mint
   module MoneyAttribute
     class Configuration
-      attr_accessor :added_currencies, :enabled_currencies, :default_currency,
+      attr_accessor :added_currencies, :default_currency,
                     :rounding_mode, :default_format
 
       def initialize
         @added_currencies = []
-        @enabled_currencies = :all
         @default_currency = 'USD'
         @rounding_mode = nil
         @default_format = nil
@@ -28,17 +27,14 @@ module Mint
 
   def self.assert_valid_currency!(currency)
     currency = Mint.currency(currency)
-    return currency if Mint.valid_currency?(currency)
+    unless currency
+      raise ArgumentError, 'Invalid currency code. Please select a registered currency'
+    end
 
-    raise ArgumentError, "Invalid currency '#{currency}'. Please select a registered currency"
+    currency
   end
 
   def self.default_currency
     @default_currency ||= Mint.assert_valid_currency!(config.default_currency)
-  end
-
-  def self.valid_currency?(currency)
-    enabled = config.enabled_currencies
-    currency && (enabled == :all || enabled.include?(currency.code))
   end
 end